Add providers

+++ src/main/java/xspleet/daggerapi/api/collections/ActionProviders.java
@@ -150,7 +150,7 @@ public class ActionProviders
                 return;
             }
             var result = amountExpression.evaluate(data);
-            data.addData(DaggerKeys.FOOD_AMOUNT, result);
+            data.addData(DaggerKeys.FOOD_AMOUNT, result.intValue());
         };
     }).addArgument(DaggerKeys.Provider.AMOUNT, DoubleExpression::create)
             .addRequiredData(DaggerKeys.FOOD_AMOUNT)
@@ -166,7 +166,7 @@ public class ActionProviders
                 return;
             }
             var result = amountExpression.evaluate(data);
-            data.addData(DaggerKeys.SATURATION_AMOUNT, result);
+            data.addData(DaggerKeys.SATURATION_AMOUNT, result.floatValue());
         };
     }).addArgument(DaggerKeys.Provider.AMOUNT, DoubleExpression::create)
             .addRequiredData(DaggerKeys.SATURATION_AMOUNT)
